fin: mui & mantine working
