+++ package.json
@@ -23,7 +23,10 @@
     "@types/react-dom": "18.2.7",
     "eslint": "8.46.0",
     "eslint-config-next": "13.4.12",
+    "gray-matter": "^4.0.3",
+    "marked": "^6.0.0",
     "next": "13.4.12",
+    "prismjs": "^1.29.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "typescript": "5.1.6"
